Refactor group related firestore actions into common package
