+++ common/firebase/common-actions.ts
@@ -1,5 +1,11 @@
-import { FirestoreTag, FirestoreTask } from '../types/firestore-types';
 import {
+  FirestoreGroup,
+  FirestoreTag,
+  FirestoreTask,
+  FirestoreUserData,
+} from '../types/firestore-types';
+import {
+  Group,
   PartialTaskMainData,
   RepeatingTaskMetadata,
   State,
@@ -29,7 +35,11 @@ type Diff = { readonly mainTaskEdits: PartialTaskMainData };
 export default class Actions {
   readonly orderManager: OrderManager;
 
-  constructor(private readonly getUserEmail: () => string, private readonly database: Database) {
+  constructor(
+    private readonly getUserEmail: () => string,
+    private readonly getUserDisplayName: () => string,
+    private readonly database: Database
+  ) {
     this.orderManager = new OrderManager(database, getUserEmail);
   }
 
@@ -79,6 +89,12 @@ export default class Actions {
         batch.commit().then(ignore);
       });
   };
+
+  /*
+   * --------------------------------------------------------------------------------
+   * Section 2: Tasks Actions
+   * --------------------------------------------------------------------------------
+   */
 
   private asyncAddTask = async (
     newTaskId: string,
@@ -235,5 +251,192 @@ export default class Actions {
       const newDiff = { ...diff, mainTaskEdits: { ...diff.mainTaskEdits, date: dateEdit } };
       this.forkTaskWithDiff(storeState, taskId, replaceDate, newDiff);
     }
+  };
+
+  /*
+   * --------------------------------------------------------------------------------
+   * Section 3: Groups Actions
+   * --------------------------------------------------------------------------------
+   */
+
+  /**
+   * Reject a group invite.
+   * @param groupID  Document ID of the group's Firestore document.
+   */
+  rejectInvite = async ({ groupInvites }: State, groupID: string): Promise<void> => {
+    const invitees = groupInvites.get(groupID)?.invitees;
+    if (!invitees) return;
+    const inviterNames = groupInvites.get(groupID)?.inviterNames;
+    if (!inviterNames) return;
+
+    // get index of inviterName to remove
+    let currIndex = 0;
+    let targetIndex = 0;
+    const email = this.getUserEmail();
+    invitees.forEach((invitee) => {
+      if (invitee === email) {
+        targetIndex = currIndex;
+      }
+      currIndex += 1;
+    });
+    // get new list of inviterNames
+    currIndex = 0;
+    const newInviterNames: string[] = [];
+    inviterNames.forEach((item) => {
+      if (currIndex !== targetIndex) {
+        newInviterNames.push(item);
+      }
+      currIndex += 1;
+    });
+    // get new list of invitees
+    const newInvitees: string[] = invitees.filter((invitee) => invitee !== email);
+    // update both invitees and inviterNames lists
+    const groupDoc = await this.database.groupsCollection().doc(groupID);
+    await groupDoc.update({ invitees: newInvitees, inviterNames: newInviterNames });
+  };
+
+  /**
+   * Get the inviter's name.
+   * @param groupID  Document ID of the group's Firestore document.
+   * @returns a string of the name of the person who sent the group invitation.
+   */
+  getInviterName = ({ groupInvites }: State, groupID: string): string => {
+    const email = this.getUserEmail(); // email of the invited person
+    const invitees = groupInvites.get(groupID)?.invitees;
+    const inviterNames = groupInvites.get(groupID)?.inviterNames;
+    if (!invitees || !inviterNames) return 'Unknown user';
+    let currIndex = 0;
+    let targetIndex = 0;
+    invitees.forEach((invitee) => {
+      if (invitee === email) {
+        targetIndex = currIndex;
+      }
+      currIndex += 1;
+    });
+    return inviterNames[targetIndex];
+  };
+
+  /**
+   * Join a group.
+   * @param groupID  Document ID of the group's Firestore document.
+   */
+  joinGroup = async (storeState: State, groupID: string): Promise<void> => {
+    const groupDoc = this.database.groupsCollection().doc(groupID);
+    const email = this.getUserEmail();
+    // Get current list of invitees (if any)
+    const invitees = await groupDoc.get().then((snapshot) => snapshot.data()?.invitees);
+    // Check if user is not in invitees list
+    if (invitees === undefined || !invitees.includes(email)) {
+      throw new Error('Invalid invitation');
+    }
+    const members = await groupDoc.get().then((snapshot) => snapshot.data()?.members);
+    // Check if user is already in the group
+    if (members === undefined || members.includes(email)) {
+      throw new Error('Invalid group members');
+    }
+    // Add user to group
+    await groupDoc.update({ members: [...members, email] });
+
+    // TODO remove this whole function after we get a Cloud Function working;
+    // the following line is a hack
+    this.rejectInvite(storeState, groupID);
+  };
+
+  /**
+   * Create a group.
+   * @param name The name of the group
+   * @param deadline Deadline for the group project
+   * @param classCode Class code for the class associated with the group
+   */
+  createGroup = (name: string, deadline: Date, classCode: string): void => {
+    const email = this.getUserEmail();
+    // creator is the only member at first
+    const newGroup: FirestoreGroup = {
+      name,
+      deadline,
+      classCode,
+      members: [email],
+      invitees: [],
+      inviterNames: [],
+    };
+    this.database.groupsCollection().doc().set(newGroup);
+  };
+
+  /**
+   * Leave a group.
+   * @param groupID Document ID of the group's Firestore document. The user calling this function must
+   *                be a member of this group.
+   */
+  leaveGroup = async ({ groups }: State, groupID: string): Promise<void> => {
+    const members = groups.get(groupID)?.members;
+    if (members === undefined) {
+      return;
+    }
+    const email = this.getUserEmail();
+    const newMembers: string[] = members.filter((m) => m !== email);
+    const groupDoc = this.database.groupsCollection().doc(groupID);
+    if (newMembers.length === 0) {
+      await groupDoc.delete();
+    } else {
+      await groupDoc.update({ members: newMembers });
+    }
+  };
+
+  updateGroup = async ({ id, ...groupInformation }: Group): Promise<void> => {
+    await this.database.groupsCollection().doc(id).update(groupInformation);
+  };
+
+  /**
+   * Send an invitation to a user to join a group.
+   * @param groupID Document ID of the group's Firestore document. The user calling this function must
+   *                be a member of this group.
+   * @param inviteeEmail The full Cornell email of the user receiving the invitation (all lowercase)
+   */
+  sendInvite = async (groupID: string, inviteeEmail: string): Promise<void> => {
+    const groupDoc = this.database.groupsCollection().doc(groupID);
+    // get current list of invites and inviterNames (if any)
+    const invitees = await groupDoc
+      .get()
+      .then((snapshot) => (snapshot.data() as FirestoreGroup)?.invitees);
+
+    const inviterNames = await groupDoc
+      .get()
+      .then((snapshot) => (snapshot.data() as FirestoreGroup)?.inviterNames);
+    this.getUserEmail();
+    const inviterName = this.getUserDisplayName();
+
+    // Add invitee to invitees if they have not already been invited
+    if (invitees === undefined) {
+      await groupDoc.update({
+        invitees: [inviteeEmail],
+        inviterNames: [inviterName],
+      });
+    } else if (!invitees.includes(inviteeEmail)) {
+      await groupDoc.update({
+        invitees: [...invitees, inviteeEmail],
+        inviterNames: [...inviterNames, inviterName],
+      });
+    }
+  };
+
+  addUserInfo = async (email: string, fullName: string, photoURL: string | null): Promise<void> => {
+    this.database.db().runTransaction(async (transaction) => {
+      const userDoc = this.database.usersCollection().doc(email);
+      const snapshot = await transaction.get(userDoc);
+
+      if (snapshot.exists) {
+        const userInfoPartial: Partial<FirestoreUserData> = {
+          name: fullName,
+          photoURL: photoURL || 'Default Photo',
+        };
+        transaction.update(userDoc, userInfoPartial);
+      } else {
+        const userInfo: FirestoreUserData = {
+          name: fullName,
+          photoURL: photoURL || 'Default Photo',
+        };
+        transaction.set(userDoc, userInfo);
+      }
+    });
   };
 }
